RestAPIの機能を追加
